shuffle to increase coverage
