--- conda_build/__main__.py
@@ -1,4 +1,0 @@
-import sys
-from conda_build.cli.main_build import main
-
-sys.exit(main())
