+++ setup.cfg
@@ -1,7 +1,7 @@
 [flake8]
 max-line-length = 100
 ignore = E122,E123,E126,E127,E128,E731
-exclude = build,conda_build/_version.py,tests,conda.recipe,.git
+exclude = build,conda_build/_version.py,tests,conda.recipe,.git,conda_build/skeletons/_example_skeleton.py
 
 [pytest]
 norecursedirs= tests/test-recipes .* *.egg* build dist conda.recipe
